refactor: 테스트 환경에서 SQLite 호환성 문제 해결을 위한 필드 타입 변경
-id 타입을 integer로 수정
- 'character varying'을 text로 수정
-nullable 가능한 필드에 true값 추가

+++ src/movies/entities/movie-provider.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 't
 import { Movie } from './movie.entity';
 @Entity('movie_providers')
 export class MovieProvider{
-  @PrimaryGeneratedColumn({ type: 'bigint' })
+  @PrimaryGeneratedColumn({ type: 'integer' })
   id: number;
 
   @Column({ type: 'bigint', name: 'movie_id' })
